LayoutTests/platform/ios-simulator/fast/events/platform-wheelevent-paging-y-in-scrolling-page-expected.txt

FAIL document.body.scrollTop should be 525. Was 0.
PASS document.body.scrollLeft is window.expectedScrollLeft
